modules/requirements.txt
drawbot-skia==0.4.6
Flat==0.3.2
fonttools==4.19.0
numpy==1.19.5
Pillow==8.1.0
pybind11==2.6.1
python-bidi==0.4.2
six==1.15.0
skia-python==87.1
uharfbuzz==0.14.0
unicodedata2==13.0.0.post2
